added file headers to frontend components, disclaimer, create export component
